# Draft: change SSE2 impl to use 8-bits Tag instead of 7