--- a/src/control/group/sse2.rs
+++ b/src/control/group/sse2.rs
@@ -73,7 +73,7 @@ impl Group {
     /// Returns a `BitMask` indicating all tags in the group which have
     /// the given value.
     #[inline]
-    pub(crate) fn match_tag(self, tag: Tag) -> BitMask {
+    pub(crate) fn match_tag(self, tag: i32) -> BitMask {
         #[allow(
             clippy::cast_possible_wrap, // tag.0: Tag as i8
             // tag: i32 as u16
@@ -83,7 +83,7 @@ impl Group {
             clippy::cast_possible_truncation
         )]
         unsafe {
-            let cmp = x86::_mm_cmpeq_epi8(self.0, x86::_mm_set1_epi8(tag.0 as i8));
+            let cmp = x86::_mm_cmpeq_epi8(self.0, x86::_mm_set1_epi32(tag));
             BitMask(x86::_mm_movemask_epi8(cmp) as u16)
         }
     }
@@ -92,7 +92,7 @@ impl Group {
     /// `EMPTY`.
     #[inline]
     pub(crate) fn match_empty(self) -> BitMask {
-        self.match_tag(Tag::EMPTY)
+        self.match_tag(Tag::EMPTY32)
     }
 
     /// Returns a `BitMask` indicating all tags in the group which are
@@ -107,15 +107,25 @@ impl Group {
             clippy::cast_possible_truncation
         )]
         unsafe {
-            // A tag is EMPTY or DELETED iff the high bit is set
-            BitMask(x86::_mm_movemask_epi8(self.0) as u16)
+            let cmp = x86::_mm_cmpgt_epi8(self.0, x86::_mm_set1_epi32(Tag::MAX_TAG32));
+            BitMask(x86::_mm_movemask_epi8(cmp) as u16)
         }
     }
 
     /// Returns a `BitMask` indicating all tags in the group which are full.
     #[inline]
     pub(crate) fn match_full(&self) -> BitMask {
-        self.match_empty_or_deleted().invert()
+        #[allow(
+            // tag: i32 as u16
+            //   note: _mm_movemask_epi8 returns a 16-bit mask in a i32, the
+            //   upper 16-bits of the i32 are zeroed:
+            clippy::cast_sign_loss,
+            clippy::cast_possible_truncation
+        )]
+        unsafe {
+            let cmp = x86::_mm_cmplt_epi8(self.0, x86::_mm_set1_epi32(Tag::DELETED32));
+            BitMask(x86::_mm_movemask_epi8(cmp) as u16)
+        }
     }
 
     /// Performs the following transformation on all tags in the group:
@@ -124,22 +134,16 @@ impl Group {
     /// - `FULL => DELETED`
     #[inline]
     pub(crate) fn convert_special_to_empty_and_full_to_deleted(self) -> Self {
-        // Map high_bit = 1 (EMPTY or DELETED) to 1111_1111
-        // and high_bit = 0 (FULL) to 1000_0000
-        //
-        // Here's this logic expanded to concrete values:
-        //   let special = 0 > tag = 1111_1111 (true) or 0000_0000 (false)
-        //   1111_1111 | 1000_0000 = 1111_1111
-        //   0000_0000 | 1000_0000 = 1000_0000
-        #[allow(
-            clippy::cast_possible_wrap, // tag: Tag::DELETED.0 as i8
-        )]
+        debug_assert_eq!(127, Tag::EMPTY.0);
+        debug_assert_eq!(126, Tag::DELETED.0);
+
+        #[allow(clippy::cast_sign_loss, clippy::cast_possible_truncation)]
         unsafe {
-            let zero = x86::_mm_setzero_si128();
-            let special = x86::_mm_cmpgt_epi8(zero, self.0);
+            let is_special = x86::_mm_cmpgt_epi8(self.0, x86::_mm_set1_epi32(Tag::MAX_TAG32));
+
             Group(x86::_mm_or_si128(
-                special,
-                x86::_mm_set1_epi8(Tag::DELETED.0 as i8),
+                x86::_mm_and_si128(is_special, x86::_mm_set1_epi32(Tag::EMPTY32)), // EMPTY if special
+                x86::_mm_set1_epi32(Tag::DELETED32), // else DELETED
             ))
         }
     }