@Amanieu 2025-09-21
Passing the benchmark results through `cargo-benchcmp` makes the differences easier to see:
```
 name                            before.txt ns/iter  after.txt ns/iter  diff ns/iter   diff %  speedup
 clone_from_large                12,657              12,717                       60    0.47%   x 1.00
 clone_from_small                131                 131                           0    0.00%   x 1.00
 clone_large                     12,859              13,012                      153    1.19%   x 0.99
 clone_small                     207                 207                           0    0.00%   x 1.00
 grow_insert_foldhash_highbits   38,844              42,258                    3,414    8.79%   x 0.92
 grow_insert_foldhash_random     42,001              42,671                      670    1.60%   x 0.98
 grow_insert_foldhash_serial     40,266              40,586                      320    0.79%   x 0.99
 grow_insert_std_highbits        71,766              72,596                      830    1.16%   x 0.99
 grow_insert_std_random          71,636              72,523                      887    1.24%   x 0.99
 grow_insert_std_serial          71,198              72,351                    1,153    1.62%   x 0.98
 insert                          13,414              13,527                      113    0.84%   x 0.99
 insert_erase_foldhash_highbits  36,820              37,095                      275    0.75%   x 0.99
 insert_erase_foldhash_random    36,681              37,033                      352    0.96%   x 0.99
 insert_erase_foldhash_serial    36,260              36,694                      434    1.20%   x 0.99
 insert_erase_std_highbits       65,048              65,396                      348    0.53%   x 0.99
 insert_erase_std_random         64,969              65,792                      823    1.27%   x 0.99
 insert_erase_std_serial         63,380              65,388                    2,008    3.17%   x 0.97
 insert_foldhash_highbits        30,413              30,409                       -4   -0.01%   x 1.00
 insert_foldhash_random          30,216              30,762                      546    1.81%   x 0.98
 insert_foldhash_serial          30,227              30,995                      768    2.54%   x 0.98
 insert_std_highbits             48,720              48,786                       66    0.14%   x 1.00
 insert_std_random               49,116              49,355                      239    0.49%   x 1.00
 insert_std_serial               49,087              48,913                     -174   -0.35%   x 1.00
 insert_unique_unchecked         10,216              11,386                    1,170   11.45%   x 0.90
 iter_foldhash_highbits          2,144               1,688                      -456  -21.27%   x 1.27
 iter_foldhash_random            2,309               1,689                      -620  -26.85%   x 1.37
 iter_foldhash_serial            2,282               1,681                      -601  -26.34%   x 1.36
 iter_std_highbits               2,355               1,662                      -693  -29.43%   x 1.42
 iter_std_random                 2,310               1,652                      -658  -28.48%   x 1.40
 iter_std_serial                 2,311               1,650                      -661  -28.60%   x 1.40
 loadfactor_lookup_14500         3,080               3,128                        48    1.56%   x 0.98
 loadfactor_lookup_16500         3,120               3,150                        30    0.96%   x 0.99
 loadfactor_lookup_18500         3,071               3,181                       110    3.58%   x 0.97
 loadfactor_lookup_20500         3,112               3,167                        55    1.77%   x 0.98
 loadfactor_lookup_22500         3,077               3,171                        94    3.05%   x 0.97
 loadfactor_lookup_24500         3,123               3,144                        21    0.67%   x 0.99
 loadfactor_lookup_26500         3,097               3,168                        71    2.29%   x 0.98
 loadfactor_lookup_28500         3,088               3,127                        39    1.26%   x 0.99
 loadfactor_lookup_fail_14500    2,653               2,295                      -358  -13.49%   x 1.16
 loadfactor_lookup_fail_16500    2,779               2,372                      -407  -14.65%   x 1.17
 loadfactor_lookup_fail_18500    2,924               2,523                      -401  -13.71%   x 1.16
 loadfactor_lookup_fail_20500    3,220               2,776                      -444  -13.79%   x 1.16
 loadfactor_lookup_fail_22500    3,775               3,293                      -482  -12.77%   x 1.15
 loadfactor_lookup_fail_24500    4,931               4,389                      -542  -10.99%   x 1.12
 loadfactor_lookup_fail_26500    7,343               6,520                      -823  -11.21%   x 1.13
 loadfactor_lookup_fail_28500    11,605              10,327                   -1,278  -11.01%   x 1.12
 lookup_fail_foldhash_highbits   5,403               4,580                      -823  -15.23%   x 1.18
 lookup_fail_foldhash_random     5,156               4,645                      -511   -9.91%   x 1.11
 lookup_fail_foldhash_serial     5,103               4,524                      -579  -11.35%   x 1.13
 lookup_fail_std_highbits        18,611              17,468                   -1,143   -6.14%   x 1.07
 lookup_fail_std_random          18,690              17,285                   -1,405   -7.52%   x 1.08
 lookup_fail_std_serial          18,349              17,238                   -1,111   -6.05%   x 1.06
 lookup_foldhash_highbits        5,714               5,598                      -116   -2.03%   x 1.02
 lookup_foldhash_random          5,661               5,554                      -107   -1.89%   x 1.02
 lookup_foldhash_serial          5,282               5,506                       224    4.24%   x 0.96
 lookup_std_highbits             19,089              18,487                     -602   -3.15%   x 1.03
 lookup_std_random               19,325              18,336                     -989   -5.12%   x 1.05
 lookup_std_serial               18,871              18,272                     -599   -3.17%   x 1.03
 rehash_in_place                 411,638             397,508                 -14,130   -3.43%   x 1.04
 set_ops_bit_and                 15,200              15,375                      175    1.15%   x 0.99
 set_ops_bit_and_assign          11,075              11,078                        3    0.03%   x 1.00
 set_ops_bit_or                  121,849             123,941                   2,092    1.72%   x 0.98
 set_ops_bit_or_assign           99,121              98,770                     -351   -0.35%   x 1.00
 set_ops_bit_xor                 127,540             122,981                  -4,559   -3.57%   x 1.04
 set_ops_bit_xor_assign          99,960              98,363                   -1,597   -1.60%   x 1.02
 set_ops_sub_assign_large_small  98,311              98,177                     -134   -0.14%   x 1.00
 set_ops_sub_assign_small_large  11,274              11,438                      164    1.45%   x 0.99
 set_ops_sub_large_small         127,703             121,118                  -6,585   -5.16%   x 1.05
 set_ops_sub_small_large         1,592               1,566                       -26   -1.63%   x 1.02
```